requirements.txt
streamlit>=1.22.0
langgraph==0.0.22
langchain==0.1.12
langchain-openai>=0.0.1
python-dotenv>=0.19.0